pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "uvicorn[standard]>=0.43.0",
]
